+++ lib/rubocop/cop/style/percent_q_literals.rb
@@ -29,8 +29,7 @@ module RuboCop
         def check(node, msg)
           # Report offense only if changing case doesn't change semantics,
           # i.e., if the string would become dynamic or has special characters.
-          return if node.children !=
-                    ProcessedSource.new(corrected(node.source)).ast.children
+          return if node.children != parse(corrected(node.source)).ast.children
 
           add_offense(node, :begin, msg)
         end