+++ lib/rubocop/error.rb
@@ -5,6 +5,7 @@ module RuboCop
   # When this exception is raised, it means that RuboCop is unable to perform
   # a requested action (probably due to misconfiguration) and must stop
   # immediately, rather than carrying on
-  class Error < StandardError
-  end
+  class Error < StandardError; end
+
+  class ValidationError < Error; end
 end