+++ spec/rubocop/cli_spec.rb
@@ -1422,4 +1422,53 @@ describe RuboCop::CLI, :isolated_environment do
       end
     end
   end
+
+  describe 'configuration of target Ruby versions' do
+    context 'when configured with an unknown version' do
+      it 'fails with an error message' do
+        create_file('.rubocop.yml', ['AllCops:',
+                                     '  TargetRubyVersion: 2.4'])
+        expect(cli.run([])).to eq(1)
+        expect($stderr.string.strip).to match(
+          /\AError: Unknown Ruby version 2.4 found in `TargetRubyVersion`/)
+        expect($stderr.string.strip).to match(
+          /Known versions: 1.9, 2.0, 2.1, 2.2, 2.3/)
+      end
+    end
+
+    context 'when set to 1.9 and Style/OptionHash is enabled' do
+      it 'fails with an error message' do
+        create_file('example1.rb', "puts 'hello'")
+        create_file('.rubocop.yml', ['AllCops:',
+                                     '  TargetRubyVersion: 1.9',
+                                     'Style/OptionHash:',
+                                     '  Enabled: true'])
+        expect(cli.run(['example1.rb'])).to eq(1)
+        expect($stderr.string.strip).to eq(
+          ['Error: The `Style/OptionHash` cop is only compatible with Ruby ' \
+           '2.0 and up, but the target Ruby version for your project is 1.9.',
+           'Please disable this cop or adjust the `TargetRubyVersion` ' \
+           'parameter in your configuration.'].join("\n"))
+      end
+    end
+
+    context 'when set to 1.9 and Style/SymbolArray is using percent style' do
+      it 'fails with an error message' do
+        create_file('example1.rb', "puts 'hello'")
+        create_file('.rubocop.yml', ['AllCops:',
+                                     '  TargetRubyVersion: 1.9',
+                                     'Style/SymbolArray:',
+                                     '  EnforcedStyle: percent',
+                                     '  Enabled: true'])
+        expect(cli.run(['example1.rb'])).to eq(1)
+        expect($stderr.string.strip).to eq(
+          ['Error: The default `percent` style for the `Style/SymbolArray` ' \
+           'cop is only compatible with Ruby 2.0 and up, but the target Ruby' \
+           ' version for your project is 1.9.',
+           'Please either disable this cop, configure it to use `array` ' \
+           'style, or adjust the `TargetRubyVersion` parameter in your ' \
+           'configuration.'].join("\n"))
+      end
+    end
+  end
 end