+++ spec/rubocop/cop/util_spec.rb
@@ -91,11 +91,13 @@ describe RuboCop::Cop::Util do
   end
 
   # Test compatibility with Range#size in Ruby 2.0.
-  describe '#numeric_range_size', ruby: 2 do
-    [1..1, 1...1, 1..2, 1...2, 1..3, 1...3, 1..-1, 1...-1].each do |range|
-      context "with range #{range}" do
-        subject { described_class.numeric_range_size(range) }
-        it { is_expected.to eq(range.size) }
+  describe '#numeric_range_size' do
+    if RUBY_VERSION >= '2'
+      [1..1, 1...1, 1..2, 1...2, 1..3, 1...3, 1..-1, 1...-1].each do |range|
+        context "with range #{range}" do
+          subject { described_class.numeric_range_size(range) }
+          it { is_expected.to eq(range.size) }
+        end
       end
     end
   end